requirements_dev.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
codegen==1.0
colorama==0.4.4
distlib==0.3.1
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.7
django-environ==0.4.5
django-extensions==3.1.1
django-heroku==0.3.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
patterns==0.3
psycopg2-binary==2.8.6
pylint==2.6.2
python-decouple==3.4
pytz==2021.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
virtualenv==20.4.2
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.12.1
